added the db_process.py file updated the page heading of html files
