+++ templates/pricipal_bill_print_template.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>PRINT BILL</title>
+        <title>EMAIL SENT::PRINT BILL</title>
         <style>
         @media print {
                       #printbtn  {
