+++ templates/principal_bill_input_template.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>ENTER SCHOOL BILLING DETAILS</title>
+        <title>ENTER BILLING DETAILS FOR PRINCIPAL</title>
         <style>
          h1 {
            text-align: center;
