+++ templates/principal_bill_output_template.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>CONFIRM PRINCIPAL BILLING DETAILS</title>
+        <title>CONFIRM BILLING DETAILS FOR PRINCIPAL</title>
         <style>
         
          h1 {
