+++ templates/student_invoice_print_template.html
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>PRINT STUDENT INVOICE</title>
+        <title>EMAIL SENT::PRINT STUDENT INVOICE</title>
         <style>
         @media print {
                       #printbtn  {
